implement likes interactions functionality
