+++ src/index.html
@@ -15,7 +15,6 @@
         <div class="movie-logo">
             <a href="#">Bingers Logo</a>
         </div>
-
         <ul class="list-group-nav">
             <li class="list-items"><a href="">My Watch List</a></li>
             <li class="list-items"><a href="">Latest Shows</a></li>
@@ -23,143 +22,7 @@
         </ul>
     </nav>
     
-    <div class="watchlist">
-        <!-- <div class="card-image">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image1">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image2">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image3">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image4">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image5">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image6">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image7">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image8">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image9">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image10">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image11">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image12">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image13">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div>
-        <div class="card-image14">
-            <div class="title">
-                <p>Blacklist</p>
-                <i class="fa fa-gittip"></i>
-            </div>
-            <div class="comments">
-                Comments
-            </div>
-        </div> -->
-    </div>
+    <div class="watchlist"></div>
     <footer>
         <div class="series-footer">
             <p>Created by Collins & Sharon under MIT License</p>
