+++ src/modules/showitem.js
@@ -1,6 +1,33 @@
-// const apiBase = 'https://api.tvmaze.com/search/shows?q=girls';
-// const getTvShow = async () => {
-//   const response = await fetch(apiBase).json();
-//   return response;
-// };
-// getTvShow();
+import { fetchList, postLikes } from './showlist.js';
+// Loading API images in the DOM
+const showSeriesList = async () => {
+  const seriesEntries = await fetchList();
+  const displaySeries = document.querySelector('.watchlist');
+  const displayItem = document.createElement('div');
+  displayItem.classList.add('card-image1');
+  for (let entry = 50; entry < 70; entry += 1) {
+    displayItem.innerHTML += `
+      <div class="card-image" id= ${seriesEntries[entry].id}>
+      <img class="show-image" src="${seriesEntries[entry].image.medium}" alt="${seriesEntries[entry].name}">
+      <p class="title">${seriesEntries[entry].name}  <i class="fa fa-heart"></i><span class="display" data-index="${seriesEntries[entry].index}" data-id="${seriesEntries[entry].id}"></span></p>
+      <button type="submit" class="comments">Comments</button> 
+      </div> 
+    `;
+    displaySeries.append(displayItem);
+  }
+  const displayLikes = document.querySelectorAll('.fa-heart');
+  displayLikes.forEach((like) => {
+    like.addEventListener('click', (e) => {
+      postLikes(e.target.parentNode.parentNode.id);
+      e.target.nextSibling.textContent = parseInt(e.target.nextSibling.textContent, 10) + 1;
+    });
+  });
+  const likeItems = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/IgKDd8EwQi9Xf3fgeQlF/likes');
+  const itemList = await likeItems.json();
+  itemList.forEach((item) => {
+    const likeItem = document.querySelector(`[data-id = '${item.item_id}']`);
+    likeItem.innerHTML = `${item.likes}`;
+  });
+};
+
+export default showSeriesList;
